# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JudgeAI - Court Judgment Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  UPLOADER: "UPLOADER",
  REVIEWER: "REVIEWER",
  VIEWER: "VIEWER",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  password: string;
  role: RegisterBodyRole;
  department?: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  department?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type JudgmentStatus =
  (typeof JudgmentStatus)[keyof typeof JudgmentStatus];

export const JudgmentStatus = {
  PENDING: "PENDING",
  PROCESSING: "PROCESSING",
  EXTRACTED: "EXTRACTED",
  VERIFIED: "VERIFIED",
  REJECTED: "REJECTED",
  EXTRACTION_FAILED: "EXTRACTION_FAILED",
} as const;

export interface Judgment {
  id: number;
  /** @nullable */
  caseNumber?: string | null;
  /** @nullable */
  courtName?: string | null;
  pdfPath: string;
  uploadedBy: number;
  uploadedAt: string;
  status: JudgmentStatus;
}

export interface JudgmentListResponse {
  judgments: Judgment[];
  total: number;
  page: number;
  limit: number;
}

export type ExtractionDataKeyDirectionsItem = { [key: string]: unknown };

export type ExtractionDataTimelinesItem = { [key: string]: unknown };

export interface ExtractionData {
  id: number;
  judgmentId: number;
  /** @nullable */
  caseTitle?: string | null;
  /** @nullable */
  caseNumber?: string | null;
  /** @nullable */
  courtName?: string | null;
  /** @nullable */
  dateOfOrder?: string | null;
  bench?: string[];
  /** @nullable */
  petitionerName?: string | null;
  /** @nullable */
  petitionerAdvocate?: string | null;
  /** @nullable */
  respondentName?: string | null;
  /** @nullable */
  respondentAdvocate?: string | null;
  keyDirections?: ExtractionDataKeyDirectionsItem[];
  timelines?: ExtractionDataTimelinesItem[];
  relevantActs?: string[];
  /** @nullable */
  summaryText?: string | null;
  /** @nullable */
  aiConfidenceScore?: number | null;
  /** @nullable */
  extractionModel?: string | null;
  /** @nullable */
  extractedAt?: string | null;
}

export type ActionPlanDataPriorityLevel =
  (typeof ActionPlanDataPriorityLevel)[keyof typeof ActionPlanDataPriorityLevel];

export const ActionPlanDataPriorityLevel = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
  CRITICAL: "CRITICAL",
} as const;

export type ActionPlanDataActionItemsItem = { [key: string]: unknown };

export interface ActionPlanData {
  id: number;
  judgmentId: number;
  complianceRequired: boolean;
  appealConsideration: boolean;
  /** @nullable */
  appealLimitationDays?: number | null;
  /** @nullable */
  appealDeadline?: string | null;
  responsibleDepartments?: string[];
  priorityLevel: ActionPlanDataPriorityLevel;
  actionItems?: ActionPlanDataActionItemsItem[];
  /** @nullable */
  natureOfAction?: string | null;
  /** @nullable */
  aiRationale?: string | null;
  generatedAt: string;
}

export interface JudgmentDetail {
  judgment: Judgment;
  extraction?: ExtractionData;
  actionPlan?: ActionPlanData;
  uploadedByUser?: User;
}

export interface VerifyQueueItem {
  id: number;
  /** @nullable */
  caseNumber?: string | null;
  /** @nullable */
  courtName?: string | null;
  uploadedAt: string;
  status: string;
  /** @nullable */
  aiConfidenceScore?: number | null;
  /** @nullable */
  caseTitle?: string | null;
  /** @nullable */
  uploadedByName?: string | null;
}

export interface VerifyDetail {
  judgment: Judgment;
  extraction?: ExtractionData;
  actionPlan?: ActionPlanData;
  pdfUrl: string;
}

export type EditVerifyBodyEditedExtraction = { [key: string]: unknown };

export type EditVerifyBodyEditedActionPlan = { [key: string]: unknown };

export interface EditVerifyBody {
  editedExtraction: EditVerifyBodyEditedExtraction;
  editedActionPlan: EditVerifyBodyEditedActionPlan;
  reviewerNotes?: string;
}

export interface RejectVerifyBody {
  reviewerNotes: string;
}

export type VerificationRecordStatus =
  (typeof VerificationRecordStatus)[keyof typeof VerificationRecordStatus];

export const VerificationRecordStatus = {
  APPROVED: "APPROVED",
  EDITED: "EDITED",
  REJECTED: "REJECTED",
} as const;

export interface VerificationRecord {
  id: number;
  judgmentId: number;
  reviewedBy: number;
  reviewedAt: string;
  status: VerificationRecordStatus;
  /** @nullable */
  reviewerNotes?: string | null;
}

export type DashboardEntryPriorityLevel =
  (typeof DashboardEntryPriorityLevel)[keyof typeof DashboardEntryPriorityLevel];

export const DashboardEntryPriorityLevel = {
  LOW: "LOW",
  MEDIUM: "MEDIUM",
  HIGH: "HIGH",
  CRITICAL: "CRITICAL",
} as const;

export type DashboardEntryKeyActionsItem = { [key: string]: unknown };

export type DashboardEntryImportantDatesItem = { [key: string]: unknown };

export type DashboardEntryStatus =
  (typeof DashboardEntryStatus)[keyof typeof DashboardEntryStatus];

export const DashboardEntryStatus = {
  ACTIVE: "ACTIVE",
  COMPLIED: "COMPLIED",
  APPEALED: "APPEALED",
  CLOSED: "CLOSED",
} as const;

export interface DashboardEntry {
  id: number;
  verificationId: number;
  judgmentId: number;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  caseNumber?: string | null;
  /** @nullable */
  caseTitle?: string | null;
  priorityLevel: DashboardEntryPriorityLevel;
  complianceRequired: boolean;
  appealConsidering: boolean;
  keyActions?: DashboardEntryKeyActionsItem[];
  importantDates?: DashboardEntryImportantDatesItem[];
  status: DashboardEntryStatus;
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  courtName?: string | null;
  /** @nullable */
  dateOfOrder?: string | null;
  /** @nullable */
  appealDeadline?: string | null;
}

export interface DashboardEntriesResponse {
  entries: DashboardEntry[];
  total: number;
  page: number;
  limit: number;
}

export type DashboardStatsByPriority = { [key: string]: unknown };

export type DashboardStatsByDepartment = { [key: string]: unknown };

export type DashboardStatsByStatus = { [key: string]: unknown };

export type DashboardStatsUpcomingDeadlinesItem = { [key: string]: unknown };

export type DashboardStatsRecentActivityItem = { [key: string]: unknown };

export interface DashboardStats {
  total: number;
  byPriority: DashboardStatsByPriority;
  byDepartment: DashboardStatsByDepartment;
  byStatus: DashboardStatsByStatus;
  upcomingDeadlines: DashboardStatsUpcomingDeadlinesItem[];
  pendingVerification: number;
  processingCount: number;
  recentActivity: DashboardStatsRecentActivityItem[];
}

export type UpdateEntryStatusBodyStatus =
  (typeof UpdateEntryStatusBodyStatus)[keyof typeof UpdateEntryStatusBodyStatus];

export const UpdateEntryStatusBodyStatus = {
  ACTIVE: "ACTIVE",
  COMPLIED: "COMPLIED",
  APPEALED: "APPEALED",
  CLOSED: "CLOSED",
} as const;

export interface UpdateEntryStatusBody {
  status: UpdateEntryStatusBodyStatus;
}

export type ListJudgmentsParams = {
  status?: string;
  page?: number;
  limit?: number;
};

export type GetDashboardEntriesParams = {
  department?: string;
  priority?: string;
  status?: string;
  dateFrom?: string;
  dateTo?: string;
  search?: string;
  page?: number;
  limit?: number;
};